+++ public/_locales/en/messages.json
@@ -1752,6 +1752,9 @@
       }
     }
   },
+  "shortcutSwitchAccountOverlayNavigation": {
+    "message": "TAB or ↑↓ to navigate, ⏎ to select,\nESC to dismiss"
+  },
   "deployNew": {
     "message": "Deploy new"
   },
