+++ src/app/hooks/use-account-select-shortcut.ts
@@ -1,4 +1,4 @@
-import { useCallback } from 'react';
+import { useCallback, useState } from 'react';
 
 import { useKeyboardShortcut } from './use-keyboard-shortcut';
 
@@ -9,34 +9,29 @@ const ACCOUNT_SELECT_HOTKEY = {
 
 const ESC_KEY = 'Escape';
 
-export const useAccountSelectShortcut = (
-  isPopperOpened: boolean,
-  setIsPopperOpened: (v: boolean) => void,
-  togglePopperOpened: EmptyFn
-) => {
+export const useAccountSelectShortcut = () => {
+  const [opened, setOpened] = useState(false);
+
   const handleEscPress = useCallback(
     (e: KeyboardEvent) => {
       if (e.key !== ESC_KEY) return;
 
       e.preventDefault();
 
-      if (isPopperOpened) {
-        setIsPopperOpened(false);
+      if (opened) {
+        setOpened(false);
       }
     },
-    [isPopperOpened, setIsPopperOpened]
+    [opened]
   );
 
-  const handleShortcutPress = useCallback(
-    (e: KeyboardEvent) => {
-      if (e.key !== ACCOUNT_SELECT_HOTKEY.key) return;
+  const handleShortcutPress = useCallback((e: KeyboardEvent) => {
+    if (e.key !== ACCOUNT_SELECT_HOTKEY.key) return;
 
-      e.preventDefault();
+    e.preventDefault();
 
-      togglePopperOpened();
-    },
-    [togglePopperOpened]
-  );
+    setOpened(prev => !prev);
+  }, []);
 
   useKeyboardShortcut({
     handler: handleEscPress
@@ -46,4 +41,6 @@ export const useAccountSelectShortcut = (
     handler: handleShortcutPress,
     modifierKey: ACCOUNT_SELECT_HOTKEY.modifierKey
   });
+
+  return { opened, setOpened };
 };
